# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface Profile {
  userId: string;
  username: string;
  /** Full Italian study-path string (e.g. "Liceo Scientifico"). Null before onboarding. */
  level?: string | null;
  wallet: number;
  streak: number;
  inviteCode: string;
  avatarObjectPath?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface UpsertProfileRequest {
  /**
     * @minLength 2
     * @maxLength 32
     */
  username: string;
  email?: string;
}

export interface UpdateLevelRequest {
  /**
     * Full Italian study-path string (e.g. "Liceo Scientifico").
     * @minLength 1
     */
  level: string;
}

/**
 * Readiness of the extracted study text. Raw extracted text is never exposed. 'ready' means usable for quizzes/flashcards.
 */
export type MaterialExtractionStatus = typeof MaterialExtractionStatus[keyof typeof MaterialExtractionStatus];


export const MaterialExtractionStatus = {
  ready: 'ready',
  unsupported: 'unsupported',
  failed: 'failed',
  pending: 'pending',
} as const;

export interface Material {
  id: string;
  ownerId: string;
  title: string;
  description?: string | null;
  contentType: string;
  objectPath: string;
  /** File size in bytes. */
  size?: number | null;
  groupId?: string | null;
  /** Readiness of the extracted study text. Raw extracted text is never exposed. 'ready' means usable for quizzes/flashcards. */
  extractionStatus: MaterialExtractionStatus;
  /** Safe Italian readiness message (never the extracted text). */
  extractionMessage: string;
  /** True when extractionStatus is 'ready'. */
  isStudyReady: boolean;
  createdAt: string;
}

export interface FinalizeMaterialRequest {
  /** @minLength 1 */
  title: string;
  description?: string;
  contentType: string;
  objectPath: string;
  /**
     * File size in bytes.
     * @minimum 1
     */
  size: number;
  groupId?: string;
}

export interface StudyGroup {
  id: string;
  name: string;
  description?: string | null;
  ownerId: string;
  coverObjectPath?: string | null;
  memberCount?: number;
  createdAt: string;
}

export interface CreateGroupRequest {
  /**
     * @minLength 1
     * @maxLength 64
     */
  name: string;
  description?: string;
  coverObjectPath?: string;
}

export interface QuizQuestion {
  question: string;
  /**
     * @minItems 4
     * @maxItems 4
     */
  options: string[];
}

export type StartQuizSessionRequestTotalQuestions = typeof StartQuizSessionRequestTotalQuestions[keyof typeof StartQuizSessionRequestTotalQuestions];


export const StartQuizSessionRequestTotalQuestions = {
  NUMBER_10: 10,
  NUMBER_20: 20,
  NUMBER_30: 30,
} as const;

export interface StartQuizSessionRequest {
  /**
     * Owned material ids to combine into one unified quiz.
     * @minItems 1
     * @maxItems 20
     */
  materialIds: string[];
  totalQuestions: StartQuizSessionRequestTotalQuestions;
}

export interface QuizSessionResponse {
  sessionId: string;
  questions: QuizQuestion[];
  expiresAt: string;
}

export interface CompleteQuizSessionRequest {
  /** Array of selected option indices (null = unanswered). Length must equal totalQuestions. */
  answers: (number | null)[];
  /** @minLength 1 */
  idempotencyKey: string;
}

export interface QuizAttempt {
  id: string;
  userId: string;
  materialId: string;
  score: number;
  totalQuestions: number;
  earnedCoins?: number;
  createdAt: string;
}

export interface FlashcardsRequest {
  /**
     * Owned, study-ready material ids.
     * @minItems 1
     * @maxItems 20
     */
  materialIds: string[];
}

export interface Flashcard {
  front: string;
  back: string;
  materialTitle: string;
}

export interface FlashcardsResponse {
  flashcards: Flashcard[];
}

export interface OwnedShopItem {
  id: string;
  userId: string;
  itemId: string;
  itemType: string;
  equipped: boolean;
  purchasedAt: string;
}

export interface BuyShopItemRequest {
  itemId: string;
}

export interface EquipShopItemRequest {
  ownedItemId: string;
}

export interface Ticket {
  id: string;
  userId: string;
  subject: string;
  /** Ticket category (e.g. "bug", "domanda", "altro"). */
  category: string;
  message: string;
  status: string;
  createdAt: string;
}

export interface CreateTicketRequest {
  /**
     * @minLength 1
     * @maxLength 128
     */
  subject: string;
  /**
     * Ticket category (e.g. "bug", "domanda", "altro").
     * @minLength 1
     */
  category: string;
  /** @minLength 1 */
  message: string;
}

export interface LeaderboardEntry {
  username: string;
  wallet: number;
}

export interface FriendEntry {
  username: string;
  /** Full Italian study-path string. Null before onboarding. */
  level?: string | null;
}

export interface InviteSummary {
  inviteCode: string;
  friends: FriendEntry[];
}

export interface UseInviteCodeRequest {
  /**
     * @minLength 6
     * @maxLength 6
     */
  code: string;
}