# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List materials owned by current user
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: finalizeMaterial
      tags: [materials]
      summary: Finalize material metadata after successful upload
      description: |
        Verifies a pending upload row owned by current user exists and is unexpired,
        confirms the object exists in storage with matching metadata/size,
        sets private ACL, inserts material record, and consumes the pending row.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinalizeMaterialRequest"
      responses:
        "201":
          description: Material created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden (pending row not owned by current user or groupId not owned)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Pending upload not found or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Upload already finalized or object already added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /materials/{materialId}:
    delete:
      operationId: deleteMaterial
      tags: [materials]
      summary: Delete a material (owner only) and remove underlying object
      parameters:
        - name: materialId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List study groups the current user is a member of
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyGroup"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a study group (also finalizes group metadata after upload)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyGroup"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /groups/{groupId}/join:
    post:
      operationId: joinGroup
      tags: [groups]
      summary: Join a study group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Joined group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyGroup"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Material:
      type: object
      required:
        - id
        - ownerId
        - title
        - contentType
        - objectPath
        - extractionStatus
        - extractionMessage
        - isStudyReady
        - createdAt
      properties:
        id:
          type: string
        ownerId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        contentType:
          type: string
        objectPath:
          type: string
        size:
          type: integer
          nullable: true
          description: File size in bytes.
        groupId:
          type: string
          nullable: true
        extractionStatus:
          type: string
          enum: [ready, unsupported, failed, pending]
          description: >
            Readiness of the extracted study text. Raw extracted text is never
            exposed. 'ready' means usable for quizzes/flashcards.
        extractionMessage:
          type: string
          description: Safe Italian readiness message (never the extracted text).
        isStudyReady:
          type: boolean
          description: True when extractionStatus is 'ready'.
        createdAt:
          type: string
          format: date-time
    FinalizeMaterialRequest:
      type: object
      required: [title, contentType, objectPath, size]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        contentType:
          type: string
        objectPath:
          type: string
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        groupId:
          type: string
    StudyGroup:
      type: object
      required: [id, name, ownerId, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        ownerId:
          type: string
        coverObjectPath:
          type: string
          nullable: true
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    CreateGroupRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        description:
          type: string
        coverObjectPath:
          type: string
...